Overwork of php code.
- Removed parts from ahbsd.Zahlensysteme.php to single files
- added a class diagram
